+++ Controller/AbstractController.php
@@ -15,12 +15,12 @@ use Psr\Log\LoggerInterface;
 use Symfony\Component\DependencyInjection\Container;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 use Symfony\Component\HttpFoundation\Session\SessionInterface;
-use Symfony\Component\HttpKernel\Kernel;
 use Symfony\Component\Routing\RouterInterface;
 use Symfony\Component\Security\Core\User\User;
 use Symfony\Component\Translation\TranslatorInterface;
 use WBW\Bundle\CoreBundle\Component\BaseEvent;
 use WBW\Bundle\CoreBundle\Event\NotificationEvent;
+use WBW\Bundle\CoreBundle\EventDispatcher\EventDispatcherHelper;
 use WBW\Bundle\CoreBundle\EventListener\KernelEventListener;
 use WBW\Bundle\CoreBundle\Exception\BadUserRoleException;
 use WBW\Bundle\CoreBundle\Helper\FormHelper;
@@ -44,17 +44,7 @@ abstract class AbstractController extends BaseController {
      * @return BaseEvent|null Returns the event in case of success, null otherwise.
      */
     protected function dispatchEvent($eventName, BaseEvent $event) {
-
-        $eventDispatcher = $this->getEventDispatcher();
-        if (null === $eventDispatcher || false === $eventDispatcher->hasListeners($eventName)) {
-            return null;
-        }
-
-        if (Kernel::VERSION_ID < 40300) {
-            return $eventDispatcher->dispatch($eventName, $event);
-        }
-
-        return $eventDispatcher->dispatch($event, $eventName);
+        return EventDispatcherHelper::dispatch($this->getEventDispatcher(), $eventName, $event);
     }
 
     /**