+++ Tests/Controller/AbstractControllerTest.php
@@ -237,22 +237,12 @@ class AbstractControllerTest extends AbstractTestCase {
      */
     public function testNotify() {
 
-        // Set a dispatch function.
-        $dispatchFunction = function(BaseEvent $event, $eventName) {
-            return $event;
-        };
-        if (Kernel::VERSION_ID < 40300) {
-            $dispatchFunction = function($eventName, BaseEvent $event) {
-                return $event;
-            };
-        }
-
         // Set a Notification mock.
         $notification = $this->getMockBuilder(NotificationInterface::class)->getMock();
 
         // Set the Event dispatcher mock.
         $this->eventDispatcher->expects($this->any())->method("hasListeners")->willReturn(true);
-        $this->eventDispatcher->expects($this->any())->method("dispatch")->willReturnCallback($dispatchFunction);
+        $this->eventDispatcher->expects($this->any())->method("dispatch")->willReturnCallback(AbstractTestCase::getEventDispatcherDispatchFunction());
 
         $obj = new TestAbstractController();
         $obj->setContainer($this->containerBuilder);