+++ Tests/Helper/FormHelperTest.php
@@ -14,7 +14,6 @@ namespace WBW\Bundle\CoreBundle\Tests\Helper;
 use Doctrine\Common\Collections\ArrayCollection;
 use Exception;
 use InvalidArgumentException;
-use WBW\Bundle\CoreBundle\Component\BaseEvent;
 use WBW\Bundle\CoreBundle\Exception\RedirectResponseException;
 use WBW\Bundle\CoreBundle\Helper\FormHelper;
 use WBW\Bundle\CoreBundle\Tests\AbstractTestCase;
@@ -89,9 +88,7 @@ class FormHelperTest extends AbstractTestCase {
 
         // Set the Event dispatcher mock.
         $this->eventDispatcher->expects($this->any())->method("hasListeners")->willReturn(true);
-        $this->eventDispatcher->expects($this->any())->method("dispatch")->willReturnCallback(function($eventName, BaseEvent $event) {
-            return $event;
-        });
+        $this->eventDispatcher->expects($this->any())->method("dispatch")->willReturnCallback(AbstractTestCase::getEventDispatcherDispatchFunction());
 
         $obj = new FormHelper($this->objectManager, $this->eventDispatcher);